+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_futures.dart
@@ -171,12 +171,14 @@ class FuturesAdapter extends OfflineFirstAdapter<Futures> {
   Future<void> afterSave(instance, {provider, repository}) async {
     if (instance.primaryKey != null) {
       await Future.wait<int>(instance.futureAssocs?.map((s) async {
-        final id = (await s)?.primaryKey ??
-            await provider?.upsert<Assoc>((await s), repository: repository);
-        return await provider?.rawInsert(
-            'INSERT OR IGNORE INTO `_brick_Futures_future_assocs` (`l_Futures_brick_id`, `f_Assoc_brick_id`) VALUES (?, ?)',
-            [instance.primaryKey, id]);
-      }) ?? []);
+            final id = (await s)?.primaryKey ??
+                await provider?.upsert<Assoc>((await s),
+                    repository: repository);
+            return await provider?.rawInsert(
+                'INSERT OR IGNORE INTO `_brick_Futures_future_assocs` (`l_Futures_brick_id`, `f_Assoc_brick_id`) VALUES (?, ?)',
+                [instance.primaryKey, id]);
+          }) ??
+          []);
     }
   }