+++ packages/brick_sqlite_generators/lib/src/sqlite_serdes_generator.dart
@@ -10,8 +10,10 @@ import 'sqlite_fields.dart';
 
 abstract class SqliteSerdesGenerator<_Model extends SqliteModel>
     extends SerdesGenerator<Sqlite, _Model> {
+  @override
   final providerName = 'Sqlite';
 
+  @override
   final String repositoryName;
 
   SqliteSerdesGenerator(