+++ packages/brick_sqlite_generators/test/migration_generator/test_create_index.dart
@@ -1,0 +1,16 @@
+import 'package:brick_sqlite_abstract/db.dart';
+
+const version = 1;
+
+const up = [
+  InsertTable('demo'),
+  InsertColumn('name', Column.varchar, onTable: 'demo'),
+  CreateIndex(columns: ['name'], onTable: 'demo'),
+];
+
+const down = <MigrationCommand>[];
+
+@Migratable(version: '$version', up: up, down: down)
+class Migration1 extends Migration {
+  const Migration1() : super(version: version, up: up, down: down);
+}