+++ packages/brick_sqlite_abstract/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## Unreleased
 
+## 0.0.9+1
+
+* Escape tables created with reserved names in the index statement. For example: `CREATE INDEX IF NOT EXISTS index_Group_on_id \`Group\`(\`id\`)`
+
+## 0.0.9
+
 * Remove `Migration.wrapInTransaction` static method. This was a misleading and unused method.
+* Add `index:` option in `@Sqlite` for creating an index on a column.
 
 ## 0.0.8+1